gtk3: compile basic schemas before packaging
Given that:

1. GtkFileChooser and GtkColorChooser use GSettings for storing some
   internal states [0];
2. schemas need to be compiled before being usable by GSettings [1];
3. a (relatively) recent commit in GLib makes GSettings crash when the
   schema is not found [2]

a pristine installation of a GTK+3 build will crash whenever a widget
that implements GtkFileChooser or GtkColorChooser is called. To
reproduce this issue, just run gtk3-demo.exe with wine and select the
"Pickers" demo.

To solve this issue the basic schemas *must* be compiled before
packaging. The native glib-compile-schemas binary is used to avoid
including an hardcoded version of gschemas.compiled: the compiled format
seems to be portable across different platforms (tested with wine with
WINEARCH set to either win32 and win64).

[0] http://git.gnome.org/browse/gtk+/tree/gtk/gtkfilechooserdefault.c?h=gtk-3-6#n6021
[1] http://developer.gnome.org/gio/2.32/glib-compile-schemas.html
[2] http://git.gnome.org/browse/glib/commit/?id=3106391694408877ebf6e8451146c5ac5d7bb017